add some filters

+++ app/Http/Controllers/MajorController.php
@@ -40,6 +40,10 @@ class MajorController extends Controller
 
     public function destroy(Major $major)
     {
+        // بررسی اینکه آیا دانش‌آموزی به این رشته متصل است
+        if ($major->students()->exists()) {
+            return back()->with('error', 'این رشته قابل حذف نیست، زیرا دانش‌آموزانی به آن متصل هستند.');
+        }
         $major->delete();
         return back()->with('success', 'رشته حذف شد.');
     }
